+++ packages/ui_flutter/test/public_api_test.dart
@@ -192,6 +192,7 @@ const publicTypes = <Type>[
   TRTabDropDetails,
   TRTabsDragConfiguration,
   TRTabsTab,
+  TRTabsWidth,
   TRText,
   TRTextAlign,
   TRTextColor,